requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
openai==1.3.5
pinecone-client==2.2.4
neo4j==5.14.1
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
Pillow==10.1.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
asyncio==3.4.3
